server: the reshard (C3) - physical semantic re-partition, validated hard
SKEG-set reshard: train the router, then per shard stream cursor-batched
moves of every live row whose semantic owner is elsewhere - payload
included, vset-then-vdel so a crash duplicates and never loses, the
vsearch merge now dedups by id (needed for exactly that window, and for
the coming boundary overlap), and a move whose id the owner map already
places at its destination only deletes the stale source copy (a
concurrent post-router write wins).

Point ops survive the new placement: an id-to-owner map (rebuilt from
the shards' live ids, lazily on first use after open - correctness
cannot depend on a caller remembering an init step) keeps vget/vdel
O(1); routed vset deletes a relocating id's old copy before writing the
new one. The reopen check covers EVERY id after a spot-check once
passed on a 50% hash-fallback coincidence.

+++ crates/skeg-platform/src/process.rs
@@ -28,6 +28,7 @@ pub fn rss_bytes() -> u64 {
         let mut count = libc::MACH_TASK_BASIC_INFO_COUNT;
         // SAFETY: task_info fills `info` up to `count` natural_t words for
         // the current task; both pointers are valid for the call.
+        #[allow(deprecated)] // mach_task_self: the mach2 crate is not worth a dep for one call
         let kr = unsafe {
             libc::task_info(
                 libc::mach_task_self(),
@@ -59,7 +60,8 @@ pub fn cpu_seconds() -> f64 {
     }
     // SAFETY: rc == 0 means the struct was written.
     let ru = unsafe { ru.assume_init() };
-    let tv = |t: libc::timeval| t.tv_sec as f64 + f64::from(t.tv_usec as i32) / 1e6;
+    #[allow(clippy::cast_precision_loss)] // sub-second fields, exact in f64
+    let tv = |t: libc::timeval| t.tv_sec as f64 + t.tv_usec as f64 / 1e6;
     tv(ru.ru_utime) + tv(ru.ru_stime)
 }
